Add image handling support and refactor models for serialization in external tools

+++ src/main/java/org/dataverse/marketplace/controller/api/ExternalToolController.java
@@ -9,10 +9,12 @@ import java.util.List;
 import org.dataverse.marketplace.model.*;
 import org.dataverse.marketplace.openapi.annotations.ExternalToolsAPIDocs;
 import org.dataverse.marketplace.payload.*;
+import org.dataverse.marketplace.security.ApplicationRoles;
 import org.dataverse.marketplace.service.ExternalToolService;
 import org.dataverse.marketplace.service.ResourceStorageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -34,18 +36,13 @@ public class ExternalToolController {
     @ExternalToolsAPIDocs.ExternalToolsListDoc
     public ResponseEntity<?> getAllTools() {
 
-        List<ExternalTool> tools = externalToolService.getAllTools();
-        ArrayList<ExternalToolDTO> toolDTOs = new ArrayList<>();
-        for (ExternalTool tool : tools) {
-            toolDTOs.add(new ExternalToolDTO(tool));
-        }
-
-        return ResponseEntity.ok(toolDTOs);
+        return ResponseEntity.ok(externalToolService.getAllTools());
     }
 
     /**
      * Method to add a new external tool
      */
+    @PreAuthorize(ApplicationRoles.ADMIN_ROLE)
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     @ExternalToolsAPIDocs.AddExternalToolsRequestDoc
     public ResponseEntity<?> addNewTool(@Valid AddToolRequest addToolRequest) {
@@ -53,9 +50,10 @@ public class ExternalToolController {
         try {
             return ResponseEntity.ok(externalToolService.addTool(addToolRequest));
         } catch (IOException e) {
-            ServerMessageResponse messageResponse = new ServerMessageResponse(HttpStatus.INTERNAL_SERVER_ERROR,
-                    "Error adding tool",
-                    e.getMessage());
+            ServerMessageResponse messageResponse 
+                = new ServerMessageResponse(HttpStatus.INTERNAL_SERVER_ERROR,
+                                            "Error adding tool",
+                                            e.getMessage());
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(messageResponse);
         }
     }
@@ -72,37 +70,50 @@ public class ExternalToolController {
     }
 
    
+    /**
+     * Method to retrieve the images of an external tool.
+     */
     @GetMapping("/{toolId}/images")
     @ExternalToolsAPIDocs.GetToolImagesDoc
     public ResponseEntity<?> getToolImages(@PathVariable("toolId") Integer toolId) {
 
         ExternalTool tool = externalToolService.getToolById(toolId);
 
-        List<Long> imagesResourceId = new ArrayList<>();
+        if(tool == null){
+            ServerMessageResponse messageResponse = new ServerMessageResponse(HttpStatus.NOT_FOUND,
+                    "Resource not found",
+                    String.format("The requested external tool with ID %d was not found.", toolId));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messageResponse);
+        }
+
+        List<MarketplaceItemImageDTO> imagesResourceId = new ArrayList<>();
         for (MarketplaceItemImage image : tool.getImages()) {
-            imagesResourceId.add(image.getImageStoredResourceId());
+            imagesResourceId.add(new MarketplaceItemImageDTO(image));
         }
         
         return ResponseEntity.ok(imagesResourceId);
     }
 
-    
+    /**
+     * Method to add images to an external tool.
+     */
+    @PreAuthorize(ApplicationRoles.ADMIN_ROLE)
     @PostMapping(path = "/{toolId}/images", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     @ExternalToolsAPIDocs.AddToolImagesDoc
     public ResponseEntity<?> addToolImages(
             @PathVariable("toolId") Integer toolId, 
             @RequestBody List<MultipartFile> images) throws IOException {
 
-        ExternalTool tool = new ExternalTool();
-        tool.setId(toolId);
+        ExternalTool tool = externalToolService.getToolById(toolId);
 
-        List<MarketplaceItemImage> mktImages = externalToolService.addItemImages(tool, images);
-
-        ArrayList<Long> imagesResourceId = new ArrayList<>();
-        
-        for(MarketplaceItemImage img : mktImages){
-            imagesResourceId.add(img.getImageStoredResourceId());
+        if(tool == null){
+            ServerMessageResponse messageResponse = new ServerMessageResponse(HttpStatus.NOT_FOUND,
+                    "Resource not found",
+                    String.format("The requested external tool with ID %d was not found.", toolId));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messageResponse);
         }
+
+        externalToolService.addItemImages(tool, images);
 
         ServerMessageResponse messageResponse = new ServerMessageResponse(HttpStatus.OK,
                 "Images added",
@@ -111,15 +122,27 @@ public class ExternalToolController {
         return ResponseEntity.ok(messageResponse);
     }
 
+    /**
+     * Method to delete an image from an external tool.
+     */
+    @PreAuthorize(ApplicationRoles.ADMIN_ROLE)
     @DeleteMapping("/{toolId}/images/{imageId}")
     @ExternalToolsAPIDocs.DeleteToolImageDoc
     public ResponseEntity<?> deleteToolImage(
             @PathVariable("toolId") Integer toolId, 
             @PathVariable("imageId") Integer imageId) {
 
+        MarketplaceItemImage image =  externalToolService.getItemImage(imageId, toolId);
+
+        if(image == null){
+            ServerMessageResponse messageResponse = new ServerMessageResponse(HttpStatus.NOT_FOUND,
+                    "Resource not found",
+                    String.format("The requested image with ID %d was not found.", imageId));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messageResponse);
+        }
+
         try {
 
-            MarketplaceItemImage image =  externalToolService.getItemImage(imageId, toolId);
             resourceStorageService.deleteResourceContent(image.getImageStoredResourceId()); 
             externalToolService.deleteToolImage(image);
             return ResponseEntity.ok(new ServerMessageResponse(HttpStatus.OK,
@@ -127,6 +150,7 @@ public class ExternalToolController {
                     String.format("The image with ID %d was deleted.", imageId)));
 
         } catch (IOException e) {
+            
             ServerMessageResponse messageResponse = new ServerMessageResponse(HttpStatus.INTERNAL_SERVER_ERROR,
                     "Error deleting image",
                     e.getMessage());
